fully working sample, still unthrottled

+++ server/index.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
+import expressPromise from 'express-promise';
 import * as es from './elastic';
 
 let app = express();
@@ -10,16 +11,15 @@ const port = 3000;
 app.use(express.static('dist'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
-
+app.use(expressPromise());
 
 app.get('/', function (req, res) {
-
 
 });
 
 app.get('/suggest/:input', function (req, res, next) {
   es.getSuggestions(req.params.input)
-  	.then((result) => { res.json(result) });
+  	.then((result) => res.json(result) );
 });
 
 /* POST document to be indexed */
@@ -28,20 +28,27 @@ app.post('/addBook', function (req, res, next) {
 	es.addDocument({
 		title: req.body.title,
 		content: req.body.content
-	}).then(() => {
-		console.log('added')
-		return res.status(200).json({
-			"message":"all ok"
-		});
-	});
+	})
+	.then(data => res.status(200).json(data))
+	.catch(e => res.status(500));
+
 });
 
 app.get('/removeIndex',(req, res) => {
-	es.deleteIndex().then(()=>{
-		return res.status(200).json({
-			message: 'deleted ;)'
-		});
+	es.indexExists().then(function(exists){
+		if (!exists) {
+			return res.status(200).json({
+				message: "index does not exist"
+			})
+		} else {
+			es.deleteIndex().then(()=>{
+				res.status(200).json({
+					message: 'deleted'
+				});
+			});
+		}
 	});
+
 });
 
 app.get('/createIndex',(req, res) => {
